Make the pi user optional

+++ scripts/root/50_disable_initial_setup.sh
@@ -1,4 +1,6 @@
 #!/bin/bash
 
-echo "Disabling initial-setup..."
-rm "$MNTDIR/etc/systemd/system/multi-user.target.wants/initial-setup-text.service"
+if [[ $HEADLESS -ne 0 ]]; then
+    echo "Disabling initial-setup..."
+    rm "$MNTDIR/etc/systemd/system/multi-user.target.wants/initial-setup-text.service"
+fi
